fix: count를 할 때 Long으로 받는 것으로 변경 - JPA의 count계열은 type 안전성의 이유로 Long으로 가져오기 때문에 service에서 type casting을 하도록 변경

+++ src/main/java/doore/document/application/convenience/DocumentConvenience.java
@@ -12,7 +12,7 @@ public class DocumentConvenience {
 
     private final DocumentRepository documentRepository;
 
-    public int countByGroupId(final Long groupId) {
+    public Long countByGroupId(final Long groupId) {
         return documentRepository.countByGroupId(groupId);
     }
 }
